feature: 日志数据预处理部分采用单独worker处理 issue #1326
