+++ src/frontend/devops-log/src/virtualScroll.vue
@@ -50,6 +50,8 @@
 
 <script>
     import language from './locale'
+    // eslint-disable-next-line
+    const DataWorker = require('worker-loader!./dataWorker.js')
 
     export default {
         props: {
@@ -96,7 +98,8 @@
                 curHoverIndex: -1,
                 hasCompleteInit: false,
                 errMessage: language('日志内容为空'),
-                isLogErr: false
+                isLogErr: false,
+                dataWorker: {}
             }
         },
 
@@ -110,6 +113,7 @@
             this.setVisWidth()
             this.initEvent()
             this.initWorker()
+            this.initAssistWorker()
         },
 
         beforeDestroy () {
@@ -118,10 +122,18 @@
             document.removeEventListener('mouseup', this.moveEnd)
             window.removeEventListener('resize', this.resize)
             window.removeEventListener('keydown', this.quickHorizontalMove)
+            this.dataWorker.terminate()
         },
 
         methods: {
             language,
+
+            initAssistWorker () {
+                this.dataWorker = new DataWorker()
+                const dataChannel = new MessageChannel()
+                this.dataWorker.postMessage({ type: 'init', dataPort: dataChannel.port1 }, [dataChannel.port1])
+                this.worker.postMessage({ type: 'initAssistWorker', dataPort: dataChannel.port2 }, [dataChannel.port2])
+            },
 
             clearSelection () {
                 window.getSelection().removeAllRanges()
@@ -319,11 +331,8 @@
                     const data = event.data
                     if (data.id !== this.id) return
                     switch (data.type) {
-                        case 'completeInit':
-                            this.freshDataScrollBottom(data)
-                            this.hasCompleteInit = true
-                            break
                         case 'completeAdd':
+                            this.hasCompleteInit = true
                             const lastIndexData = this.indexList[this.indexList.length - 1] || { listIndex: 0 }
                             if (this.totalNumber - lastIndexData.listIndex <= 3) {
                                 this.freshDataScrollBottom(data)
@@ -391,8 +400,7 @@
             },
 
             addLogData (list) {
-                const type = this.hasCompleteInit ? 'addListData' : 'initLog'
-                const postData = { type, list, mainWidth: this.mainWidth, id: this.id }
+                const postData = { type: 'addListData', list, mainWidth: this.mainWidth, id: this.id }
                 this.worker.postMessage(postData)
             },
