+++ src/frontend/devops-pipeline/src/components/ExecDetail/log/worker.js
@@ -1,60 +1,19 @@
-import ansiParse from './ansiParse'
-
-const colorList = [
-    { key: '##[command]', color: 'rgba(146,166,202,1)' },
-    { key: '##[info]', color: 'rgba(127,202,84,1)' },
-    { key: '##[warning]', color: 'rgba(246,222,84,1)' },
-    { key: '##[error]', color: 'rgba(247,49,49,1)' },
-    { key: '##[debug]', color: 'rgba(99,176,106,1)' }
-]
-
-const reg = (() => {
-    const colors = colorList.map((color) => {
-        let key = color.key
-        key = key.replace(/\[|\]/gi, '\\$&')
-        return `(${key})`
-    })
-    return new RegExp(`${colors.join('|')}`, 'gi')
-})()
-
-function handleColor (val) {
-    const parseRes = ansiParse(val) || [{ message: '' }]
-    const res = { message: '' }
-    parseRes.forEach((item) => {
-        res.message += item.message
-        if (!res.color && item.color) res.color = item.color
-    })
-    
-    const currentColor = colorList.find(color => String(val).startsWith(color.key))
-    if (currentColor) {
-        res.color = currentColor.color
-        res.message = String(res.message).replace(reg, '')
-    }
-    if (res.color) res.fontWeight = 600
-    return res
-}
-
 const allListData = {}
 let curListData = []
 const allTagList = {}
-let curTagList = []
-const allMainWidth = {}
-const allMainWordNum = {}
-const allFoldLineNum = {}
-const allRepeatLineNum = {}
 const currentSearch = {
     index: 0,
     val: ''
 }
 let curId
 let searchRes
+let dataPort
 
 onmessage = function (e) {
     const data = e.data
     const type = data.type
     curId = data.id
     curListData = allListData[curId]
-    curTagList = allTagList[curId]
 
     switch (type) {
         case 'initStatus':
@@ -62,23 +21,13 @@ onmessage = function (e) {
             pluginList.forEach((curId) => {
                 allListData[curId] = []
                 allTagList[curId] = []
-                allMainWidth[curId] = 0
-                allMainWordNum[curId] = 0
-                allRepeatLineNum[curId] = -1
-                allFoldLineNum[curId] = 0
             })
             break
-        case 'initLog':
-            allMainWidth[curId] = data.mainWidth - 90
-            allMainWordNum[curId] = Math.floor(allMainWidth[curId] / 6.8)
-            addListData(data)
-            postMessage({ type: 'completeInit', number: curListData.length, id: curId })
+        case 'initAssistWorker':
+            initAssistWorker(data)
             break
         case 'addListData':
-            allMainWidth[curId] = data.mainWidth - 90
-            allMainWordNum[curId] = Math.floor(allMainWidth[curId] / 6.8)
-            addListData(data)
-            postMessage({ type: 'completeAdd', number: curListData.length, id: curId })
+            dataPort.postMessage({ type: 'addListData', list: data.list, mainWidth: data.mainWidth, curId })
             break
         case 'wheelGetData':
             getListData(data)
@@ -110,14 +59,40 @@ onmessage = function (e) {
         case 'resetData':
             const resetList = [
                 { data: allListData, default: [] },
-                { data: allTagList, default: [] },
-                { data: allMainWidth, default: 0 },
-                { data: allMainWordNum, default: 0 },
-                { data: allRepeatLineNum, default: -1 },
-                { data: allFoldLineNum, default: 0 }
+                { data: allTagList, default: [] }
             ]
             resetData(resetList)
+            dataPort.postMessage({ type: 'resetData' })
             break
+    }
+}
+
+function initAssistWorker (data) {
+    dataPort = data.dataPort
+    dataPort.onmessage = (e) => {
+        const data = e.data
+        const type = data.type
+
+        switch (type) {
+            case 'complateHandleData':
+                const tempId = data.curId
+                data.list.forEach((newItem) => {
+                    if (newItem.message.startsWith('##[group]')) {
+                        newItem.message = newItem.message.replace('##[group]', '')
+                        allTagList[tempId].push(newItem)
+                    }
+    
+                    if (newItem.message.startsWith('##[endgroup]') && allTagList[tempId].length) {
+                        newItem.message = newItem.message.replace('##[endgroup]', '')
+                        const linkItem = allTagList[tempId].pop()
+                        linkItem.endIndex = newItem.realIndex
+                        linkItem.children = []
+                    }
+                    allListData[tempId].push(newItem)
+                })
+                postMessage({ type: 'completeAdd', number: allListData[tempId].length, id: tempId })
+                break
+        }
     }
 }
 
@@ -198,142 +173,15 @@ function foldListData ({ startIndex }) {
                 totalNum = 0
             }
             const subList = curListData.splice(realIndex + 1, currentNum)
-            allFoldLineNum[curId] -= subList.length
             currentItem.children = currentItem.children.concat(subList)
         }
     } else {
         for (let index = 0; index < currentItem.children.length;) {
             const someList = currentItem.children.slice(index, index + 10000)
             curListData.splice(realIndex + 1 + index, 0, ...someList)
-            allFoldLineNum[curId] += someList.length
             index += 10000
         }
         currentItem.children = []
-    }
-}
-
-function addListData ({ list }) {
-    list.forEach((item) => {
-        const newItemArr = (item.message || '').split(/\r\n|\n/)
-        newItemArr.forEach((val) => {
-            const { message, color } = handleColor(val || '')
-            const splitTextArr = splitText(message)
-            splitTextArr.forEach((message, i) => {
-                const currentIndex = curListData.length
-                const newItem = {
-                    message,
-                    color,
-                    isNewLine: i > 0 ? (allRepeatLineNum[curId]++, true) : false,
-                    showIndex: currentIndex - allRepeatLineNum[curId] - allFoldLineNum[curId],
-                    realIndex: currentIndex + allFoldLineNum[curId],
-                    timestamp: item.timestamp
-                }
-                if (message.startsWith('##[group]')) {
-                    newItem.message = newItem.message.replace('##[group]', '')
-                    curTagList.push(newItem)
-                }
-
-                if (message.startsWith('##[endgroup]') && curTagList.length) {
-                    newItem.message = newItem.message.replace('##[endgroup]', '')
-                    const linkItem = curTagList.pop()
-                    linkItem.endIndex = currentIndex
-                    linkItem.children = []
-                }
-
-                curListData.push(newItem)
-            })
-        })
-    })
-}
-
-function splitText (message) {
-    let tempMes = ''
-    const totalWidth = getTextWidth(message)
-    const mesRes = []
-    if (totalWidth < allMainWidth[curId]) {
-        mesRes.push(message)
-    } else {
-        const regex = /<a[^>]+?href=["']?([^"']+)["']?[^>]*>([^<]+)<\/a>/gi
-        const aList = []
-        let tempA = null
-        let currentIndex = 0
-
-        while ((tempA = regex.exec(message)) != null) {
-            aList.push({
-                content: tempA[0],
-                href: tempA[1],
-                text: tempA[2],
-                startIndex: tempA.index
-            })
-        }
-        if (aList.length) message = message.replace(regex, '$2')
-
-        while (message !== '') {
-            [tempMes, message] = splitByChar(message)
-            // a标签单独处理
-            aList.forEach((x) => {
-                if (x.startIndex <= currentIndex + tempMes.length && x.startIndex >= currentIndex) {
-                    const curStartIndex = x.startIndex - currentIndex
-                    const curLength = x.text.length
-                    const diffDis = curStartIndex + curLength - tempMes.length
-                    if (diffDis > 0) {
-                        message = tempMes.slice(curStartIndex) + message
-                        tempMes = tempMes.slice(0, curStartIndex)
-                    } else {
-                        tempMes = (tempMes.slice(0, curStartIndex) + x.content + tempMes.slice(curStartIndex + curLength))
-                    }
-                }
-            })
-
-            currentIndex += tempMes.length
-            mesRes.push(tempMes)
-        }
-    }
-    return mesRes
-}
-
-function splitByChar (message) {
-    let tempMes = message.slice(0, allMainWordNum[curId])
-    message = message.slice(allMainWordNum[curId])
-    let tempWidth = getTextWidth(tempMes)
-    while (tempWidth > allMainWidth[curId] || (allMainWidth[curId] - tempWidth > 15 && message !== '')) {
-        if (tempWidth > allMainWidth[curId]) {
-            message = tempMes.slice(-1) + message
-            tempMes = tempMes.slice(0, -1)
-            tempWidth = getTextWidth(tempMes)
-        } else {
-            tempMes += message.slice(0, 1)
-            message = message.slice(1)
-            tempWidth = getTextWidth(tempMes)
-        }
-    }
-    return [tempMes, message]
-}
-
-let getTextWidth
-if (self.OffscreenCanvas) {
-    const canvas = new OffscreenCanvas(100, 1)
-    const context = canvas.getContext('2d')
-    context.font = 'normal 12px Consolas, "Courier New", monospace'
-    getTextWidth = (text) => {
-        const metrics = context.measureText(text)
-        return metrics.width
-    }
-} else {
-    // 兼容safari
-    getTextWidth = (text) => {
-        let res = 0
-        let disNum = 0
-        for (let i = 0, len = text.length; i < len; i++) {
-            if (/[\u4e00-\u9fa5]/.test(text[i])) {
-                res += 12
-            } else {
-                res += 8
-                disNum++
-            }
-        }
-        res -= (disNum - 1)
-        return res + 20
     }
 }
